[Tom] #150 update the generator to fail on duplicate fields
The FIX specification states:

"A tag number (field) should only appear in a message once. If it
appears more than once in the message it should be considered an error
with the specification document. The error should be pointed out to
the FIX Global Technical Committee."

Failing the code generation informs the user the FIX spec they are
using is "invalid"; attempting to use the generated codecs may
result in unexpected behaviour