Improve validate method generated by message decoders.
Currently, the validate method on a message only checks the children
fields of a message and none of it's descendants. This means the user
of a message decode would have to call validate whilst iterating through
a repeating group.

This change updates the validate method on a message decode. The validate
method will now iterate through any repeating groups (if there are any)
and call validate on each entry in that group.