Change order field validation occurs.
At the moment the decoders check if a required field has been seen
before checking to see if any unknown fields have been visited. This can
be misleading in the case of an unknown field being seen before a
required field as the validation failure will refer to a missing
required field where the real failure is due to an unknown field being
present.