Don't use "allFields" for flow control.
If a Message decoder comes across an unknown field, it should exit if
validation is off or skip over the field if validation is on. For all
other aggregates, stop trying to decode the field and return the it's
parent. Checking if a field is in the FIX schema (unrelated to the
Message) for determine whether to skip it or not does not make sense