[Tom/Andrey] call reset on all repeating group entries.
Currently, reset is only being called on the first entry of a repeating group, which can result in corrupt/invalid data when future messages are parsed. If you receive a message with 3 entries in a repeating group with all fields populated, and then parse another message (after calling reset) with 3 entries in the repeating group but those entries do not have optional fields set, then you will send up with a corrupt message as the optional fields are still set from the first message that was parsed.

This commit iterates over entries in a repeating group and resets all of them.