Fix edge case for unknown nested repeating groups.
This is similar to the fix for "alreadySeenFields" (see 5ee8a43e5e56ddf352ddcce1d34ed0a54d6a23b3).

If an unknown nested group with multiple entries (more than 2) is reached whilst decoding a repeating group entry, early termination of decoding the repeating group will occur. This is due to the tag values for the unknown nested repeating group being added to the "seenFields" set, which is used to determine when to move onto the next repeating group entry. For an unknown field, we should just remove it from the seenFields set as the decoder didn't actually handle ti