[Tom] Don't generated nested component fields.
When generating an interface for a component, do not generate the fields
for any nested components on that interface. The fields do not belong to
that interface but the nested interface. Instead have the component
extend the interface of the nested component. This follows the same
pattern as how groups and messages handle components (they implement
their interface)

checkpoint - fix for why some group classes are missing. Interface extension was happening in the wrong place and breaks for nested groups