[Tom] Fix infinite loop  when decoding malformed FIX messages.
If a decoder receives a malformed FIX message (not complete tag / value
combo or missing pipes) then the decoders will get stuck in an infinite
loop trying to find the whole message.

This scenario does not occur if using the whole artio library due to
how message delimitation works in the TCP layer.