[Tom/Richard] Carry on decoding a message for field that does not belong at the top/message level
When decoding a message that has an unexpected tag at the top level of the message that is defined
in a repeating group, we should not stop decoding the message (if we are not failing for unknow field)
but skip that tag and continue with the next one.

There are examples of venues that break the FIX spec and will send tags defined for a repeating group
(in their spec!) outside of the group. In these cases we just want to skip the tag and continue as usual.

If CODEC_REJECT_UNKNOWN_FIELD_ENABLED is enabled then the message will stop decoding