fix #452 by making DecimalFloat constants read only
The purpose of this change is to avoid subtle bugs when the seemingly constant value is redefined.

The change tried to conform to the following constraints:

- maintain as much backward compatibility as possible with a minimal or no change required by the library user

- do not adversely affect the performance

When it comes to the backward compatibility,
the only required change on the client side (if any) should
be in places where constant such as ZERO or NAN were assigned or passed  to a function that expected DecimalFloat.

The client can either:
- change expected DecimalFloat type to ReadOnlyDecimalFloat
- ask for a mutableCopy()

There should be no other changes required (comparison, equality etc. maintains backward compatibility and works fine with both types interchangeably)

If there are other changes required, e.g. invocation of a mutating method such as setValue no longer compiles, it means that the upgrade detected a potential error in the client's codebase that tried to modify a constant value. This is the very reason the change was introduced.