Fixed multiple connections after connection timeouts
Let's say you have an initiator and your counter-party is dropping SYN packets which causes connection timeouts.
If you try to initiate a few times, the `Reply<Session>`s will all time out. Then, when the counter-party starts
properly accepting connections, Artio might establish multiple connections for that session:

```
[FIX_CONNECTION]Connecting to localhost:10000 from library 2142515358
[FIX_CONNECTION]Initiating session 1 with correlationId 6816641249315627913 from library 2142515358
[FIX_CONNECTION]Initiating session 1 with correlationId 6816641249315627912 from library 2142515358
[FIX_CONNECTION]Initiating session 1 with correlationId 6816641249315627914 from library 2142515358
```

The problem is that the reply timeout comes from the `SessionConfiguration.timeoutInMs()`
and not from the underlying socket, which is still open and the OS is trying to reconnect
(described here: https://blog.cloudflare.com/when-tcp-sockets-refuse-to-die/#synsent).

I think the problem is simply that `DefaultTcpChannelSupplier` doesn't close the channel in `stopConnecting`.

You'll probably want to delete my test, because it's manual and hence ignored, but it does show the issue.