Reset buffer references in string-like fields in generated encoders
The generated encoders contain different ways of setting string-like fields.  Some methods take buffers owned by the application, e.g. passing in an AsciiSequenceView.  Other methods take data that is copied into the internal buffer, e.g. passing in a CharSequence.

If an application uses both styles, it could first pass in an AsciiSequenceView (which would point to some kind of input buffer), then on a later code path, pass in a CharSequence.  This will result in the encoder copying the CharSequence into the input buffer, corrupting that buffer.

This change takes a reference to the buffer created in the UnsafeBuffer (owned by the encoder) whenever the application passes in an application-owned buffer, like an AsciiSequenceView.  On reset, it resets the UnsafeBuffer to point it back to the original buffer.