Customisable cancel on disconnect behaviour (#482)
This change provides a customizable cancel on disconnect option to an
engine that will be applied in the case a login message does not provide
one. It also adds an optional timeout window along with it.

The use case if for supporting an api that always uses cancel on
disconnect and does not want to provide an option in the FIX api.