Avoid leaking offline sessions.
Relates to support ticket 18482.

When using `soleLibraryMode` and a FIX session disconnects, the engine
will keep the session instance in `LiveLibraryInfo` but mark it as
disconnected (e.g., removing the endpoint information and
`connectionId`).

When a FIX session reconnects, a new session will be added to
`LiveLibraryInfo` regardless of whether or not there is an existing
session.

In this commit, I've added a removal as part of `addSession` maintaining
an invariant that `LiveLibraryInfo` has only one session instance per
`sessionId`. I believe this is safe, but I'm not hugely familiar with
lots of the code paths. I have added some logging in case I have missed
something. There are some long standing checks around duplicate
sessions, which gives me hope.